+++ kvs-indexing/src/main/java/org/gbif/kvs/indexing/options/HBaseIndexingOptions.java
@@ -57,9 +57,9 @@ public interface HBaseIndexingOptions extends PipelineOptions {
   void setSaltedKeyBuckets(int saltedKeyBuckets);
 
   @Description("GBIF API connection time-out in seconds")
-  long getApiTimeOut();
+  Integer getApiTimeOut();
 
-  void setApiTimeOut(long apiTimeOut);
+  void setApiTimeOut(Integer apiTimeOut);
 
   @Description("Rest/HTTP client file-cache max size")
   long getRestClientCacheMaxSize();
